Create Network_Delay_Time.cpp
You are given a network of n nodes, labeled from 1) to n.
You are also given times, a list of travel times as directed
edges times [i] = (Ui, Vi, Wi), where ui is the source
node, vi is the target node, and wi is the time it takes for a signal to travel from source to target.
We will send a signal from a given node k. Return the minimum time it takes for all the n nodes to receive the signal. If it is impossible for all the n nodes to receive the signal, return -1.