Implement Floyd-Warshall algorithm for city reachability
There are n cities numbered from 0 to (n-1). Given the array
edges where edges [i] = [fromi, toi, weightil represents
a bidirectional and weighted edge between cities from and toi, and given the integer distanceThreshold.
Return the city with the smallest number of cities that are reachable through some path and whose distance is at most distanceThreshold, If there are multiple such cities, return the city with the greatest number.
Notice that the distance of a path connecting cities i and j is equal to the sum of the edges' weights along that path.